http/protobuf support (#621)
Closes #419.

Adds support for the standard protocols for metrics and traces.

Specific protocol choices win out over generic ones. There are With... functions for all three destinations, and a typed object for specifying them.

I specified the protocol down in pipelines, but didn't want to expose pipelines for constants, so I also promoted them to the top level.